Split Crashlytics in an own library
