+++ LogcatCoreLib/build.gradle
@@ -9,7 +9,7 @@ android {
         versionName "1.0.0"
 
         minSdkVersion 16
-        project.archivesBaseName = "Logcat"
+        project.archivesBaseName = "LogcatCore"
     }
 
     compileOptions {
@@ -26,6 +26,4 @@ dependencies {
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
     api 'com.google.android.material:material:1.1.0'
     api 'com.jakewharton.timber:timber:4.7.1'
-    api 'com.google.firebase:firebase-analytics:17.4.3'
-    api 'com.google.firebase:firebase-crashlytics:17.0.1'
 }
