+++ LogcatCoreLib/src/main/java/info/hannes/logcat/LogcatActivity.kt
