+++ LogcatCoreLib/src/main/java/info/hannes/logcat/LogcatFragment.kt
