+++ LogcatCoreLib/src/main/java/info/hannes/logcat/LogfileActivity.kt
