+++ LogcatCoreLib/src/main/java/info/hannes/logcat/LogfileFragment.kt
