+++ LogcatCoreLib/src/main/java/info/hannes/logcat/TabsAdapter.kt
