+++ LogcatCoreLib/src/main/java/info/hannes/logcat/base/LoadingDialog.kt
