+++ LogcatCoreLib/src/main/java/info/hannes/logcat/base/LogBaseFragment.kt
