+++ LogcatCoreLib/src/main/java/info/hannes/logcat/base/LogListAdapter.kt
