+++ LogcatCoreLib/src/main/java/info/hannes/timber/DebugTree.kt
