+++ LogcatCoreLib/src/main/java/info/hannes/timber/FileLoggingTree.kt
