+++ LogcatCoreLib/src/main/java/info/hannes/timber/TimberExtensions.kt
