+++ LogcatCoreLib/src/main/res/layout/activity_log.xml
