+++ LogcatCoreLib/src/main/res/layout/fragment_log.xml
