+++ LogcatCoreLib/src/main/res/layout/item_log.xml
