+++ LogcatCoreLib/src/main/res/menu/menu_log.xml
