+++ LogcatCoreLib/src/main/res/raw/gitlog.json
