+++ LogcatCoreLib/src/main/res/values/colors.xml
