+++ LogcatCoreLib/src/main/res/values/strings.xml
