+++ LogcatCoreLib/src/main/res/values/styles.xml
