+++ LogcatCrashlyticLib/src/main/java/info/hannes/logcat/crashlytic/BothLogsFragment.kt
@@ -1,4 +1,4 @@
-package info.hannes.logcat
+package info.hannes.logcat.crashlytic
 
 import android.annotation.SuppressLint
 import android.os.Bundle
@@ -8,6 +8,9 @@ import android.view.ViewGroup
 import android.widget.TabHost
 import androidx.fragment.app.Fragment
 import androidx.viewpager.widget.ViewPager
+import info.hannes.logcat.LogcatFragment
+import info.hannes.logcat.LogfileFragment
+import info.hannes.logcat.TabsAdapter
 import info.hannes.logcat.base.LogBaseFragment.Companion.MAIL_LOGGER
 
 /**
