+++ LogcatCrashlyticLib/src/main/res/layout/fragment_both_logs.xml
