+++ sample/build.gradle
@@ -37,7 +37,8 @@ android {
 }
 
 dependencies {
-    implementation project(':LogcatLib')
+    implementation project(':LogcatCoreLib')
+    implementation project(':LogcatCrashlyticLib')
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation "androidx.appcompat:appcompat:1.1.0"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
